+++ src/config.rs
@@ -33,7 +33,11 @@ use serde::{Deserialize, Serialize};
 use serde_json::json;
 use tracing::info;
 
-use crate::{environment::Environment, logger, Error, Result};
+use crate::{
+    controller::middleware::{remote_ip::RemoteIPConfig, secure_headers::SecureHeadersConfig},
+    environment::Environment,
+    logger, scheduler, Error, Result,
+};
 
 lazy_static! {
     static ref DEFAULT_FOLDER: PathBuf = PathBuf::from("config");
@@ -69,6 +73,8 @@ pub struct Config {
     /// accessing `ctx.config.settings`.
     #[serde(default)]
     pub settings: Option<serde_json::Value>,
+
+    pub scheduler: Option<scheduler::Config>,
 }
 
 /// Logger configuration
@@ -348,7 +354,7 @@ pub struct Workers {
 }
 
 /// Worker mode configuration
-#[derive(Clone, Default, Serialize, Deserialize, Debug)]
+#[derive(Clone, Default, Serialize, Deserialize, Debug, PartialEq, Eq)]
 pub enum WorkerMode {
     /// Workers operate asynchronously in the background, processing queued
     /// tasks. **Requires a Redis connection**.
@@ -383,8 +389,29 @@ pub struct Middlewares {
     /// Serving static assets
     #[serde(rename = "static")]
     pub static_assets: Option<StaticAssetsMiddleware>,
+    /// Sets a set of secure headers
+    pub secure_headers: Option<SecureHeadersConfig>,
+    /// Calculates a remote IP based on `X-Forwarded-For` when behind a proxy
+    pub remote_ip: Option<RemoteIPConfig>,
+    /// Configure fallback behavior when hitting a missing URL
+    pub fallback: Option<FallbackConfig>,
 }
 
+#[derive(Debug, Clone, Deserialize, Serialize)]
+pub struct FallbackConfig {
+    /// By default when enabled, returns a prebaked 404 not found page optimized
+    /// for development. For production set something else (see fields below)
+    pub enable: bool,
+    /// For the unlikely reason to return something different than `404`, you
+    /// can set it here
+    pub code: Option<u16>,
+    /// Returns content from a file pointed to by this field with a `404` status
+    /// code.
+    pub file: Option<String>,
+    /// Returns a "404 not found" with a single message string. This sets the
+    /// message.
+    pub not_found: Option<String>,
+}
 /// Static asset middleware configuration
 #[derive(Debug, Clone, Deserialize, Serialize)]
 pub struct StaticAssetsMiddleware {