+++ src/gen/templates/scaffold/html/view_show.t
@@ -4,24 +4,21 @@ to: assets/views/{{file_name}}/show.html
 skip_exists: true
 message: "{{file_name}} view was added successfully."
 ---
-<!DOCTYPE html>
-<html lang="en">
+{% raw %}{% extends "base.html" %}{% endraw %}
 
-<head>
-    <script src="https://cdn.tailwindcss.com?plugins=forms,typography,aspect-ratio,line-clamp"></script>
-</head>
+{% raw %}{% block title %}{% endraw %}
+View {{name}}: {% raw %}{{ item.id }}{% endraw %}
+{% raw %}{% endblock title %}{% endraw %}
 
-<body class="prose p-10">
-    <h1>View {{name}}: {% raw %}{{ item.id }}{% endraw %}</h1>
-    <div class="mb-10">
-     {% for column in columns -%}
-        <div>
+{% raw %}{% block content %}{% endraw %}
+<h1>View {{name}}: {% raw %}{{ item.id }}{% endraw %}</h1>
+<div class="mb-10">
+{% for column in columns -%}
+    <div>
         <label>{{column.0}}: {% raw %}{{item.{% endraw %}{{column.0}}{% raw %}}}{% endraw %}</label>
-        </div>
-    {% endfor -%}
-    <br />
-    <a href="/{{name | plural}}">Back to {{name | plural}}</a>
     </div>
-</body>
-
-</html>
+{% endfor -%}
+<br />
+<a href="/{{name | plural}}">Back to {{name | plural}}</a>
+</div>
+{% raw %}{% endblock content %}{% endraw %}